add rtmpprotocolstack class code

+++ libuv-1.8.0/include/uv-win.h
@@ -24,7 +24,8 @@
 #endif
 
 #if !defined(_SSIZE_T_) && !defined(_SSIZE_T_DEFINED)
-typedef intptr_t ssize_t;
+//typedef intptr_t ssize_t;
+typedef unsigned int ssize_t;
 # define _SSIZE_T_
 # define _SSIZE_T_DEFINED
 #endif
